Extracted code related to the screen shot.
